Example of the world
XMLs and Pictures are paired by their names. This is the XML file you
will have to process and create a world based on it. This is NOT the
final world; Your component has to be able to recognize the elements and
store them in your world, with coordinates, if necessary,
transformations.